Navigation

+++ app/app.py
@@ -1,4 +1,5 @@
 from app.movie import Movie
+from app.list import List
 import json
 
 
@@ -16,14 +17,18 @@ class App:
         result = {}
         if self.__method == 'OPTIONS':
             return result
-        movie = Movie(self.__data['name'])
-        if self.__method == 'POST':
-            result = movie.get()
-        elif self.__method == 'PUT':
-            result = movie.set(self.__data['id'])
-        elif self.__method == 'DELETE':
-            if self.__data['file']:
-                result = movie.delete()
-            else:
-                result = movie.clear()
+        if 'list' in self.__data:
+            list = List(self.__data['list'])
+            result = list.list()
+        else:
+            movie = Movie(self.__data['name'])
+            if self.__method == 'POST':
+                result = movie.get()
+            elif self.__method == 'PUT':
+                result = movie.set(self.__data['id'])
+            elif self.__method == 'DELETE':
+                if self.__data['file']:
+                    result = movie.delete()
+                else:
+                    result = movie.clear()
         return json.dumps(result)
